+++ src/main/java/com/apocfarce/minestuck_alternia/data/AlterniaEnUsLanguageProvider.java
@@ -6,9 +6,9 @@ import com.apocfarce.minestuck_alternia.Item.block.PortalBaseItem;
 import com.apocfarce.minestuck_alternia.Item.block.PortalCenterItem;
 import com.apocfarce.minestuck_alternia.Item.block.PortalCrownItem;
 import com.apocfarce.minestuck_alternia.Minestuck_alternia;
-import com.apocfarce.minestuck_alternia.Utils.CustomDamageSources;
 import com.apocfarce.minestuck_alternia.block.AlterniaBlocks;
 import com.apocfarce.minestuck_alternia.client.StartDimensionSelectionScreen;
+import com.apocfarce.minestuck_alternia.util.CustomDamageSources;
 import com.apocfarce.minestuck_alternia.world.biome.AlterniaBiomes;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.item.ItemGroup;
